+++ src/render/OGL2_0/CVertexBufferOGL2_0.cpp
@@ -33,7 +33,7 @@ CVertexBufferOGL2_0::~CVertexBufferOGL2_0()
     Destroy();
 }
 
-void CVertexBufferOGL2_0::Initialize(size_t elementSize)
+void CVertexBufferOGL2_0::Initialize(uint64_t elementSize)
 {
     if (!IsValid())
     {
@@ -84,17 +84,17 @@ bool CVertexBufferOGL2_0::IsValid() const
     return (m_Id != 0);
 }
 
-size_t CVertexBufferOGL2_0::SizeRaw() const
+uint64_t CVertexBufferOGL2_0::SizeRaw() const
 {
     return m_Buffer.size();
 }
 
-void CVertexBufferOGL2_0::ResizeRaw(size_t newSize)
+void CVertexBufferOGL2_0::ResizeRaw(uint64_t newSize)
 {
     m_Buffer.resize(newSize);
 }
 
-size_t CVertexBufferOGL2_0::ElementSize() const
+uint64_t CVertexBufferOGL2_0::ElementSize() const
 {
     return m_ElementSize;
 }
@@ -174,9 +174,9 @@ void CVertexBufferOGL2_0::Unbind()
 // Protected Methods
 // *****************************************************************************
 
-void CVertexBufferOGL2_0::ElementSize(size_t elementSize)
+void CVertexBufferOGL2_0::ElementSize(uint64_t elementSize)
 {
-    m_ElementSize = std::max<size_t>(elementSize, 1);
+    m_ElementSize = std::max<uint64_t>(elementSize, 1);
 }
 
 // *****************************************************************************
